server/requirement.txt
aioredis==1.1.0
appnope==0.1.0
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
backcall==0.1.0
channels==2.1.0
channels-redis==2.2.1
constantly==15.1.0
daphne==2.2.0
decorator==4.3.0
django==2.0.8
djangorestframework==3.8.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
msgpack==0.5.6
parso==0.3.0
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==18.4.0
txaio==2.10.0
wcwidth==0.1.7
zope.interface==4.5.0
django-environ==0.4.4
psycopg2==2.7.3.1
django-extensions==2.1.4